# [DataGrid] Filtering performance