--- a/packages/grid/x-data-grid-generator/src/services/real-data-service.ts
+++ b/packages/grid/x-data-grid-generator/src/services/real-data-service.ts
@@ -60,3 +60,44 @@ export function getRealGridData(
     });
   });
 }
+
+export function getRealGridDataSync(
+  rowLength: number,
+  columns: GridColDefGenerator[],
+): GridDemoData {
+  const rows: GridDemoData['rows'] = [];
+  const indexedValues: { [field: string]: { [value: string]: number } } = {};
+
+  for (let i = 0; i < rowLength; i += 1) {
+    const row: any = {};
+
+    for (let j = 0; j < columns.length; j += 1) {
+      const column = columns[j];
+      if (column.generateData) {
+        const context: GridDataGeneratorContext = {};
+        if (column.dataGeneratorUniquenessEnabled) {
+          let fieldValues = indexedValues[column.field];
+          if (!fieldValues) {
+            fieldValues = {};
+            indexedValues[column.field] = fieldValues;
+          }
+
+          context.values = fieldValues;
+        }
+
+        row[column.field] = column.generateData(row, context);
+      }
+    }
+
+    rows.push(row);
+  }
+
+  const columnVisibilityModel: GridColumnVisibilityModel = {};
+  columns.forEach((col) => {
+    if (col.hide) {
+      columnVisibilityModel[col.field] = false;
+    }
+  });
+
+  return { columns, rows, initialState: { columns: { columnVisibilityModel } } };
+}