--- a/packages/grid/x-data-grid-premium/src/hooks/features/rowGrouping/gridRowGroupingUtils.ts
+++ b/packages/grid/x-data-grid-premium/src/hooks/features/rowGrouping/gridRowGroupingUtils.ts
@@ -139,10 +139,8 @@ export const filterRowTreeFromGroupingColumns = (
       }
     }
 
-    if (isPassingFiltering && areAncestorsExpanded)
-      visibleRowsLookup.add(node.id);
-    if (isPassingFiltering)
-      filteredRowsLookup.add(node.id);
+    if (isPassingFiltering && areAncestorsExpanded) visibleRowsLookup.add(node.id);
+    if (isPassingFiltering) filteredRowsLookup.add(node.id);
 
     // TODO rows v6: Should we keep storing the visibility status of footer independently or rely on the group visibility in the selector ?
     if (node.type === 'group' && node.footerId != null) {