--- a/packages/grid/x-data-grid-pro/src/components/headerFiltering/GridHeaderFilterAdornment.tsx
+++ b/packages/grid/x-data-grid-pro/src/components/headerFiltering/GridHeaderFilterAdornment.tsx
@@ -103,7 +103,7 @@ GridHeaderFilterAdornment.propTypes = {
   }).isRequired,
   operators: PropTypes.arrayOf(
     PropTypes.shape({
-      getApplyFilterFn: PropTypes.func.isRequired,
+      getApplyFilterFn: PropTypes.func,
       getValueAsString: PropTypes.func,
       headerLabel: PropTypes.string,
       InputComponent: PropTypes.elementType,