--- a/packages/grid/x-data-grid-pro/src/hooks/features/treeData/gridTreeDataUtils.ts
+++ b/packages/grid/x-data-grid-pro/src/hooks/features/treeData/gridTreeDataUtils.ts
@@ -30,8 +30,8 @@ export const filterRowTreeFromTreeData = (
   params: FilterRowTreeFromTreeDataParams,
 ): Omit<GridFilterState, 'filterModel'> => {
   const { rowTree, disableChildrenFiltering, isRowMatchingFilters } = params;
-  const visibleRowsLookup: Record<GridRowId, boolean> = {};
-  const filteredRowsLookup: Record<GridRowId, boolean> = {};
+  const visibleRowsLookup = new Set<GridRowId>();
+  const filteredRowsLookup = new Set<GridRowId>();
   const filteredDescendantCountLookup: Record<GridRowId, number> = {};
 
   const filterTreeNode = (
@@ -86,13 +86,15 @@ export const filterRowTreeFromTreeData = (
       }
     }
 
-    visibleRowsLookup[node.id] = shouldPassFilters && areAncestorsExpanded;
-    filteredRowsLookup[node.id] = shouldPassFilters;
+    if (shouldPassFilters && areAncestorsExpanded)
+      visibleRowsLookup.add(node.id);
+    if (shouldPassFilters)
+      filteredRowsLookup.add(node.id);
 
     // TODO: Should we keep storing the visibility status of footer independently or rely on the group visibility in the selector ?
     if (node.type === 'group' && node.footerId != null) {
-      visibleRowsLookup[node.footerId] =
-        shouldPassFilters && areAncestorsExpanded && !!node.childrenExpanded;
+      if (shouldPassFilters && areAncestorsExpanded && !!node.childrenExpanded)
+        visibleRowsLookup.add(node.footerId);
     }
 
     if (!shouldPassFilters) {