--- a/packages/grid/x-data-grid-pro/src/hooks/features/treeData/gridTreeDataUtils.ts
+++ b/packages/grid/x-data-grid-pro/src/hooks/features/treeData/gridTreeDataUtils.ts
@@ -92,10 +92,8 @@ export const filterRowTreeFromTreeData = (
       }
     }
 
-    if (shouldPassFilters && areAncestorsExpanded)
-      visibleRowsLookup.add(node.id);
-    if (shouldPassFilters)
-      filteredRowsLookup.add(node.id);
+    if (shouldPassFilters && areAncestorsExpanded) visibleRowsLookup.add(node.id);
+    if (shouldPassFilters) filteredRowsLookup.add(node.id);
 
     // TODO: Should we keep storing the visibility status of footer independently or rely on the group visibility in the selector ?
     if (node.type === 'group' && node.footerId != null) {