--- a/packages/grid/x-data-grid/src/colDef/gridBooleanOperators.ts
+++ b/packages/grid/x-data-grid/src/colDef/gridBooleanOperators.ts
@@ -5,13 +5,13 @@ import { GridFilterOperator } from '../models/gridFilterOperator';
 export const getGridBooleanOperators = (): GridFilterOperator<any, boolean | null, any>[] => [
   {
     value: 'is',
-    getApplyFilterFn: (filterItem: GridFilterItem) => {
+    getApplyFilterFnV7: (filterItem: GridFilterItem) => {
       if (!filterItem.value) {
         return null;
       }
 
       const valueAsBoolean = filterItem.value === 'true';
-      return ({ value }): boolean => {
+      return (value, _, __): boolean => {
         return Boolean(value) === valueAsBoolean;
       };
     },