--- a/packages/grid/x-data-grid/src/colDef/gridDateOperators.ts
+++ b/packages/grid/x-data-grid/src/colDef/gridDateOperators.ts
@@ -1,7 +1,6 @@
 import { GridFilterInputDate } from '../components/panel/filterPanel/GridFilterInputDate';
 import { GridFilterItem } from '../models/gridFilterItem';
-import { GridFilterOperator } from '../models/gridFilterOperator';
-import { GridCellParams } from '../models/params/gridCellParams';
+import { GridFilterOperator, ApplyFilterV7 } from '../models/gridFilterOperator';
 
 const dateRegex = /(\d+)-(\d+)-(\d+)/;
 const dateTimeRegex = /(\d+)-(\d+)-(\d+)T(\d+):(\d+)/;
@@ -11,7 +10,7 @@ function buildApplyFilterFn(
   compareFn: (value1: number, value2: number) => boolean,
   showTime?: boolean,
   keepHours?: boolean,
-) {
+): ApplyFilterV7 | null {
   if (!filterItem.value) {
     return null;
   }
@@ -23,7 +22,7 @@ function buildApplyFilterFn(
 
   const time = new Date(year, month - 1, day, hour || 0, minute || 0).getTime();
 
-  return ({ value }: GridCellParams<any, Date, any>): boolean => {
+  return (value, _, __): boolean => {
     if (!value) {
       return false;
     }
@@ -47,7 +46,7 @@ function buildApplyFilterFn(
 export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any, Date, any>[] => [
   {
     value: 'is',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(filterItem, (value1, value2) => value1 === value2, showTime);
     },
     InputComponent: GridFilterInputDate,
@@ -55,7 +54,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'not',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(filterItem, (value1, value2) => value1 !== value2, showTime);
     },
     InputComponent: GridFilterInputDate,
@@ -63,7 +62,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'after',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(filterItem, (value1, value2) => value1 > value2, showTime);
     },
     InputComponent: GridFilterInputDate,
@@ -71,7 +70,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'onOrAfter',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(filterItem, (value1, value2) => value1 >= value2, showTime);
     },
     InputComponent: GridFilterInputDate,
@@ -79,7 +78,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'before',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(
         filterItem,
         (value1, value2) => value1 < value2,
@@ -92,7 +91,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'onOrBefore',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       return buildApplyFilterFn(filterItem, (value1, value2) => value1 <= value2, showTime);
     },
     InputComponent: GridFilterInputDate,
@@ -100,8 +99,8 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'isEmpty',
-    getApplyFilterFn: () => {
-      return ({ value }): boolean => {
+    getApplyFilterFnV7: () => {
+      return (value, _, __): boolean => {
         return value == null;
       };
     },
@@ -109,8 +108,8 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   },
   {
     value: 'isNotEmpty',
-    getApplyFilterFn: () => {
-      return ({ value }): boolean => {
+    getApplyFilterFnV7: () => {
+      return (value, _, __): boolean => {
         return value != null;
       };
     },