--- a/packages/grid/x-data-grid/src/colDef/gridDateOperators.ts
+++ b/packages/grid/x-data-grid/src/colDef/gridDateOperators.ts
@@ -1,6 +1,6 @@
 import { GridFilterInputDate } from '../components/panel/filterPanel/GridFilterInputDate';
 import { GridFilterItem } from '../models/gridFilterItem';
-import { GridFilterOperator, ApplyFilterV7 } from '../models/gridFilterOperator';
+import { GridFilterOperator, GridApplyFilterV7 } from '../models/gridFilterOperator';
 
 const dateRegex = /(\d+)-(\d+)-(\d+)/;
 const dateTimeRegex = /(\d+)-(\d+)-(\d+)T(\d+):(\d+)/;
@@ -10,7 +10,7 @@ function buildApplyFilterFn(
   compareFn: (value1: number, value2: number) => boolean,
   showTime?: boolean,
   keepHours?: boolean,
-): ApplyFilterV7 | null {
+): GridApplyFilterV7 | null {
   if (!filterItem.value) {
     return null;
   }
@@ -22,7 +22,7 @@ function buildApplyFilterFn(
 
   const time = new Date(year, month - 1, day, hour || 0, minute || 0).getTime();
 
-  return (value, _, __): boolean => {
+  return (value, _, __, ___): boolean => {
     if (!value) {
       return false;
     }
@@ -100,7 +100,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   {
     value: 'isEmpty',
     getApplyFilterFnV7: () => {
-      return (value, _, __): boolean => {
+      return (value, _, __, ___): boolean => {
         return value == null;
       };
     },
@@ -109,7 +109,7 @@ export const getGridDateOperators = (showTime?: boolean): GridFilterOperator<any
   {
     value: 'isNotEmpty',
     getApplyFilterFnV7: () => {
-      return (value, _, __): boolean => {
+      return (value, _, __, ___): boolean => {
         return value != null;
       };
     },