--- a/packages/grid/x-data-grid/src/colDef/gridSingleSelectOperators.ts
+++ b/packages/grid/x-data-grid/src/colDef/gridSingleSelectOperators.ts
@@ -13,32 +13,34 @@ const parseObjectValue = (value: unknown) => {
 export const getGridSingleSelectOperators = (): GridFilterOperator[] => [
   {
     value: 'is',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       if (filterItem.value == null || filterItem.value === '') {
         return null;
       }
-      return ({ value }): boolean => parseObjectValue(value) === parseObjectValue(filterItem.value);
+      return (value, _, __): boolean =>
+        parseObjectValue(value) === parseObjectValue(filterItem.value);
     },
     InputComponent: GridFilterInputSingleSelect,
   },
   {
     value: 'not',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       if (filterItem.value == null || filterItem.value === '') {
         return null;
       }
-      return ({ value }): boolean => parseObjectValue(value) !== parseObjectValue(filterItem.value);
+      return (value, _, __): boolean =>
+        parseObjectValue(value) !== parseObjectValue(filterItem.value);
     },
     InputComponent: GridFilterInputSingleSelect,
   },
   {
     value: 'isAnyOf',
-    getApplyFilterFn: (filterItem) => {
+    getApplyFilterFnV7: (filterItem) => {
       if (!Array.isArray(filterItem.value) || filterItem.value.length === 0) {
         return null;
       }
       const filterItemValues = filterItem.value.map(parseObjectValue);
-      return ({ value }): boolean => filterItemValues.includes(parseObjectValue(value));
+      return (value, _, __): boolean => filterItemValues.includes(parseObjectValue(value));
     },
     InputComponent: GridFilterInputMultipleSingleSelect,
   },