… packages/grid/x-data-grid/src/colDef/gridStringOperators.ts
… packages/grid/x-data-grid/src/hooks/features/columns/gridColumnsSelector.ts
… packages/grid/x-data-grid/src/hooks/features/filter/gridFilterSelector.ts
… packages/grid/x-data-grid/src/hooks/features/filter/gridFilterState.ts
… packages/grid/x-data-grid/src/hooks/features/filter/gridFilterUtils.ts
… packages/grid/x-data-grid/src/hooks/features/filter/useGridFilter.tsx
… packages/grid/x-data-grid/src/hooks/features/rows/gridRowsSelector.ts
… packages/grid/x-data-grid/src/hooks/features/rows/useGridParamsApi.ts
… packages/grid/x-data-grid/src/hooks/features/rows/useGridRows.ts
… packages/grid/x-data-grid/src/hooks/utils/useGridApiMethod.ts
… packages/grid/x-data-grid/src/internals/index.ts
… packages/grid/x-data-grid/src/models/api/gridParamsApi.ts
… packages/grid/x-data-grid/src/models/colDef/gridColDef.ts
… packages/grid/x-data-grid/src/models/gridFilterOperator.ts
… packages/grid/x-data-grid/src/utils/createSelector.ts