### Summary

This PR is a POC to demonstrate how we can improve the performance of our filtering, and other operations. It's not meant to be merged as it is, I'm just exploring ways to evolve our API & architecture to improve performance. Please add comments.

https://deploy-preview-9120--material-ui-x.netlify.app/x/react-data-grid/filtering/#header-filters

### Results

This PR improves the speed of one-column string-contains filtering by about 75%.
*(edit: the first 2 commits of this PR do, and the later commits improve to 93%)*

### Changes & observations

Our biggest cost by far is memory allocations. The amount of objects & functions (closures) being allocated to pass data around creates much more CPU work than anything else.

##### 1. Fast filters

The first change is to define some of our filter functions as "fast filters". What this means is that they only use the `cellParams.value` field of their input arguments, so we can avoid calling the expensive API function `.getCellParams()` and instead only call `.getCellValue()`.

In practice, from what I've seen, *all* of our filters could be fast filters, so we could change the API. But the filter API is public so we can't change it without a major version increase.

##### 2. Memoize filtered items

The second change is to avoid calling `getFilterCallbackFromItem()` during `passFilterLogic()`. The former is called simply to filter the model items, but it's a very expensive call (it basically recreates the filter function) and it was done for every row, so for N rows we were re-creating the filtering function N times (plus 1, for the filtering function actually used for filtering).

*Edit: I've added more changes below, see next comment for details. Each commit correspond to one change, and it's probably easier to read each commit independently than the PR change as a whole.*

### Benchmarks

The measurements were done by filtering 100,000 rows with the string `"am"`. The `Elapsed (ms)` results below are generated by wrapping the `flatFilteringMethod` with `performance.now()` timings.

<img src="https://github.com/mui/mui-x/assets/1423607/74f69bec-dbaa-48a8-9bd8-a19772f39745" width="350" />

| Before | After |
| --- | --- |
| ![Screenshot from 2023-05-25 19-52-23](https://github.com/mui/mui-x/assets/1423607/87bf44d4-71e2-4fa1-9989-49eaba396c8b) | ![Screenshot from 2023-05-25 19-53-59](https://github.com/mui/mui-x/assets/1423607/55d83ad9-30ca-4f96-bb5c-3b363dfe66a4) |

*NOTE: The results above don't use the same set of 100,000 rows (it's the `Employee` dataset, randomly generated), thus why the `Count` differs. These results are approximate but give an accurate perspective and were consistently reproducible.*