@romgrk 2023-05-26
I've added more changes, some of them quite exotic, to see how much I could push the performance. They've improved the filtering by another 10%, so 85% compared to the baseline. They're described here.

### Changes

##### 3. Non-memoized selectors

The selectors created with `createSelector()` have a cost, which become apparent when they're called in a loop like filtering does. Most of our selector don't need the memoization of `createSelector()`, so I've added `createRawSelector()` to avoid paying the memoization cost.

I recommend that we apply this change to all our existing selectors who don't need memoization (all those that don't derive their input argument).

##### 4. Reduce API methods indirection

The `useGridApiMethod` hook creates indirection by wrapping in a trampoline function. If we remove it, we avoid function calls and allocations (the rest arguments).

```
privateApiRef.current.register(visibility, {
  [methodName]: (...args: any[]) => {
    const fn = apiMethodsRef.current[methodName] as Function;
    return fn(...args);
  },
});
```

If there aren't apparent downsides to this, I would suggest that we do it.

##### 5. Avoid dynamic object props

The pattern `{ [dynamicKey]: value }` is expensive for JS engines, I've used `eval(...)` to create a function with a static `{ key: value }`. This is exotic/evil, I don't really recommend that we implement it, but it's an interesting optimization.

##### 6. Direct state access

Avoid selectors and using direct `apiRef.current.state.` access in our internal functions can improve the performance, selector functions (even unmemoized) still have a cost.

##### 7. Use `Set`

We often use simple objects for use cases where we use them for `Set` operations. Using the proper data structure improves performance, but these changes are also semver major because they're selectable by the public selectors.


### Benchmarks

<img src="https://github.com/mui/mui-x/assets/1423607/60dddca4-5946-4fb0-8d28-54f2d7a6be64" width="400" />