@romgrk 2023-05-28
Did some more profiling, I found two more substantial improvements:

##### 8. Remove API Proxy wrapper

The wrapper around the API object is expensive, in particular if it needs unwrapping. Removing it would improve our performance across all the DataGrid functions.

##### 9. Iterate rows directly

We are currently iterating row ids in `flatFilteringMethod`, which is slow and unefficient because then we need to do `dataRowIdToModelLookup[id]` to retrieve the row, and it is a huge cost if it's a huge object. If we know we are filtering all the rows, then we're much better off iterating row objects directly. Below is a comparison.

**Before** *(Builtins_KeyedLoadIC_Megamorphic is the relevant entry)*
![Screenshot from 2023-05-28 02-53-11](https://github.com/mui/mui-x/assets/1423607/9b9439f0-8f93-4485-8202-b04509493a1a)

**After**
![Screenshot from 2023-05-28 13-49-21](https://github.com/mui/mui-x/assets/1423607/a0703131-b56a-4d67-9a1c-6f2cb5fafa74)


### Final results

The improvement compared to the baseline is 93% with these last changes. I haven't added the commits here yet, I'll cleanup what I have an evaluate which commits we can merge cleanly without a semver major change.  

<img src="https://github.com/mui/mui-x/assets/1423607/34f74ad4-dfce-4a16-bebb-9498f7cb93db" width="200" />