@romgrk 2023-05-29
Before I start fixing the tests, could I have a quick overall review? I'd just like to ensure that everyone is ok with the architectural changes. I'll comment on specific issues below, and open a new thread if you see something wrong.