@oliviertassinari 2023-06-03
From what I can quickly test on 

PR: https://deploy-preview-9120--material-ui-x.netlify.app/x/react-data-grid/filtering/#header-filters vs. HEAD: https://mui.com/x/react-data-grid/filtering/header-filters/

with:

<img width="291" alt="Screenshot 2023-06-04 at 00 18 33" src="https://github.com/mui/mui-x/assets/3165635/f0a5a204-438b-491d-86c9-9b1d93fb83bb">

We are down from 500ms to 126ms. I suspect that we should reduce the debounce https://github.com/mui/mui-x/blob/8201d0c6a70ea9ed89fbe44d5c60ecca7956ee40/packages/grid/x-data-grid/src/components/toolbar/GridToolbarQuickFilter.tsx#L63-L67

it feels like a lot.