@romgrk 2023-06-12
TODO: `getDefaultFilterModel()` allocates a new value each time and is called on every iteration.